requirements.txt
python-dotenv
PyYAML
openai
google-generativeai
streamlit
faiss-cpu
numpy
sentence-transformers
